create custom hook for loading
